Job_Board/requirements.txt
requests[socks]>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pymongo>=4.6.1
fake-useragent>=1.4.0
PySocks>=1.7.1
cloudscraper>=1.2.71
aiohttp>=3.9.0
asyncio>=3.4.3
pandas>=2.1.0
tqdm>=4.66.0 